Rename "general-schema-parser.ts" to "parser.ts"
