+++ src/test/programmatic-schema/boolean.test.ts
@@ -1,4 +1,4 @@
-import { parse } from '../../general-schema-parser'
+import { parse } from '../../parse'
 import { validate } from '../../validate'
 import { boolean } from '../../programmatic-schema'
 import { check, unknownX } from '../test-utils'
